Requested changes: remove unnecessary code + queue tasks with explicit queue instead of in the workers
